Sección 13: Auth.js - Autenticación de protección de rutas

+++ 03-admin-todos/src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse } from 'next/server'
 import { bool, object, string } from 'yup';
@@ -23,9 +24,16 @@ const postSchema = object({
 })
 
 export async function POST(req: Request) {
+
+    const user = await getUserSessionServer();
+
+    if (!user){
+        return NextResponse.json('No authorizado', {status: 401});
+    }
+
     try {
         const { complete, description } = await postSchema.validate(await req.json());
-        const todo = await prisma.todo.create({data: {complete, description}});
+        const todo = await prisma.todo.create({data: {complete, description, userId: user.id}});
         return NextResponse.json(todo);
 
     } catch (error) {
@@ -35,8 +43,16 @@ export async function POST(req: Request) {
 
 
 export async function DELETE(req: Request){
+
+    const user = await getUserSessionServer();
+
+    if (!user){
+        return NextResponse.json('No authorizado', {status: 401});
+    }
+
+
     try {
-        await prisma.todo.deleteMany({where: {complete: true}})
+        await prisma.todo.deleteMany({where: {complete: true, userId: user.id}})
         return NextResponse.json('Borrados');
     } catch (error){
         return NextResponse.json(error, {status: 400})
